product-review-analyzer/requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
chromadb==0.4.22
python-dotenv==1.0.0
openai==1.12.0
tiktoken==0.5.2
pandas==2.2.0